refactor(api,core): move CompetitionWithout* DTOs to Core
The four admin data-quality DTOs (CompetitionsWithoutCompetitors/Drives/
Metrics/Plays) lived under the API's Canonical data models. Move them to
a new SportsData.Core/Dtos/Competition namespace so Producer can share
them when it grows its own versions of these queries.

Also untangles a file-hygiene wart while relocating:
CompetitionWithoutMetricsDto was declared inside
CompetitionWithoutDrivesDto.cs; each DTO now has its own file.

Consumers (four query handlers, their tests, CanonicalAdminDataProvider,
AdminController) are pure using swaps — no behavioral change.
