+++ src/SportsData.Api/Application/Admin/Queries/GetCompetitionsWithoutCompetitors/GetCompetitionsWithoutCompetitorsQueryHandler.cs
@@ -1,9 +1,8 @@
 using FluentValidation.Results;
 
 using SportsData.Api.Infrastructure.Data.Canonical;
-using SportsData.Api.Infrastructure.Data.Canonical.Models;
-using SportsData.Core.Dtos.Canonical;
 using SportsData.Core.Common;
+using SportsData.Core.Dtos.Competition;
 
 namespace SportsData.Api.Application.Admin.Queries.GetCompetitionsWithoutCompetitors;
