--- src/SportsData.Api/Infrastructure/Data/Canonical/Models/CompetitionWithoutPlaysDto.cs
@@ -1,13 +1,0 @@
-using System;
-
-namespace SportsData.Api.Infrastructure.Data.Canonical.Models;
-
-public class CompetitionWithoutPlaysDto
-{
-    public Guid ContestId { get; set; }
-    public string? ContestName { get; set; }
-    public DateTime StartDateUtc { get; set; }
-    public Guid CompetitionId { get; set; }
-    public int PlayCount { get; set; }
-    public string? LastPlayText { get; set; }
-}
